Move Contracts diagnostics aspect tests to Diagnostics folder/namespace.
