+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/NotNull_Ineligible_GenericStruct.t.cs
