+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/NotNull_Ineligible_Int.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) SharpCrafters s.r.o. See the LICENSE.md file in the root directory of this repository root for details.
 
-namespace Metalama.Patterns.Contracts.AspectTests;
+namespace Metalama.Patterns.Contracts.AspectTests.Diagnostics;
 
 public class NotNull_Ineligible_Int
 {
