+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_DecimalGreaterThanDouble.t.cs
