+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_DecimalStrictlyGreaterThanDouble.t.cs
