+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_DecimalStrictlyLessThanDouble.t.cs
