+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_InvalidType_Byte.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) SharpCrafters s.r.o. See the LICENSE.md file in the root directory of this repository root for details.
 
-namespace Metalama.Patterns.Contracts.AspectTests;
+namespace Metalama.Patterns.Contracts.AspectTests.Diagnostics;
 
-public class Range_InvalidType_Byte4
+public class Range_InvalidType_Byte
 {
-    [GreaterThan( 256.0 )]
+    [GreaterThan( 256 )]
     private byte field;
 }
