+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_InvalidType_Byte2.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) SharpCrafters s.r.o. See the LICENSE.md file in the root directory of this repository root for details.
 
-namespace Metalama.Patterns.Contracts.AspectTests;
+namespace Metalama.Patterns.Contracts.AspectTests.Diagnostics;
 
 public class Range_InvalidType_Byte2
 {
