+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_InvalidType_Byte2.t.cs
