+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_InvalidType_Byte3.t.cs
