+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_LongGreaterThanDouble.t.cs
