+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_LongStrictlyGreaterThanDouble.t.cs
