+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_LongStrictlyLessThanDouble.t.cs
