+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_UlongGreaterThanDouble.t.cs
