+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_UlongStrictlyGreaterThanDouble.t.cs
