+++ src/tests/Metalama.Patterns.Contracts.AspectTests/Diagnostics/Range_UlongStrictlyLessThanUlong.t.cs
